ce_cloud_models/PaddleHub/AUDIO/linux/requirements.txt
colorama
colorlog
easydict
filelock
flask >= 1.1.0
gitpython
numpy
matplotlib
opencv-python
packaging
paddle2onnx >= 0.5.1
paddlenlp >= 2.0.0rc5
Pillow
pyyaml
pyzmq
rarfile
tqdm
visualdl >= 2.0.0
# gunicorn not support windows
gunicorn >= 19.10.0; sys_platform != "win32"

librosa
